Added a TiledCameraView class.
Added a TiledCameraView class to allow for situations where you for
instance want to do a fly-through of a scene over X amount of screens
that are next to each other. The other 3D method is currently for the
cameras to look around in a circle and content being animated in a
circle around them.